libbpf: provide libbpf API to encode BTF kind information
This can be used by BTF parsers to handle kinds they do not know about;
this is useful when the encoding libbpf is more recent than the parsing
BTF; the parser can then skip over the encoded types it does not know
about.

We use BTF to encode the BTF kinds that are known at the time of
BTF encoding; the use of basic BTF kinds (structs, arrays, base types)
to describe each kind and any associated metadata allows BTF parsing
to handle new kinds that the parser (in libbpf or the kernel) does
not know about.  These kinds will not be used, but since we know
their format they can be skipped over and the rest of the BTF can
be parsed.  This means we can encode BTF without worrying about the
kinds a BTF parser knows about, and means we can avoid using
--skip_new_kind solutions.  This is valuable, as if kernel BTF encodes
everything it can, something as simple as a libbpf package update
then unlocks that encoded information, whereas if we encode
pessimistically and drop representations of new kinds, this is not
possible.

So, in short, by carrying a representation of all the kinds encoded,
parsers can parse all of the encoded kinds, even if they cannot use
them all.

We use BTF itself to carry this representation because this approach
does not require BTF parsing to understand a new BTF header format;
BTF parsing simply sees some additional types it does not do anything
with.  However, a BTF parser that knows about the encoding of kind
information can use this information to guide parsing.

The process works by explicitly adding btf structs for each kind.
Each struct consists of a "struct __btf_type" followed by an array of
metadata structs representing the following metadata (for those kinds
that have it).  For kinds where a single metadata structure is used,
the metadata array has one element.  For kinds where the number
of metadata elements varies as per the info.vlen field, a zero-element
array is encoded.

For a given kind, we add a struct __BTF_KIND_<kind>.  For example,

struct __BTF_KIND_INT {
	struct __btf_type type;
};

For a type with one metadata element, the representation looks like
this:

struct __BTF_KIND_META_ARRAY {
	__u32 type;
	__u32 index_type;
	__u32 nelems;
};

struct __BTF_KIND_ARRAY {
	struct __btf_type type;
	struct __BTF_KIND_META_ARRAY meta[1];
};

For a type with an info.vlen-determined number of following metadata
objects, a zero-length array is used:

struct __BTF_KIND_STRUCT {
	struct __btf_type type;
	struct __BTF_KIND_META_STRUCT meta[0];
};

In order to link kind numeric kind values to the appropriate struct,
a typedef is added; for example:

typedef struct __BTF_KIND_INT __BTF_KIND_1;

When BTF parsing encounters a kind that is not known, the
typedef __BTF_KIND_<kind number> is looked up, and we find which
struct type id it points to.  So

	1 -> typedef __BTF_KIND_1 -> struct __BTF_KIND_INT

This approach is preferred, since it ensures the structs representing
BTF kinds have names which match their associated kind rather than
an opaque number.

From there, BTF parsing can look up that struct and determine
	- its basic size;
	- if it has metadata; and if so
	- how many array instances are present;
		- if 0, we know it is a vlen-determined number;
		  i.e. vlen * meta_size
		- if > 0, simply use the overall struct size;

Based upon that information, BTF parsing can proceed for such
unknown kinds, since sufficient information was provided
at encoding time to skip over them.

Note that this assumes that the above kind-related data
structures are represented in BTF _prior_ to any kinds that
are new to the parser.  It also assumes the basic kinds
required to represent kinds + metadata; base types, structs,
arrays, etc.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>