libbpf: use BTF-encoded kind information to help parse unrecognized kinds
btf__add_kinds() adds typedef/struct representations of the kinds supported
at BTF encoding time.  When decoding/parsing, we can then use information
about unrecognized kinds to skip over them.  This will be useful if the
BTF encoder encoded info using a new kind, but the parser doesn't support
it yet.

Note that only size determinations of unrecognized kinds are supported;
lookup, dedup and other features are not supported; the aim here is to
not break BTF parsing if newer kinds are encountered.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>